Fixed some lidar stuff

+++ src/rover_ws/rover_lidar/main.py
@@ -16,8 +16,8 @@ def callback(msg):
         lenr = len(msg.ranges)
         startidx = int(lenr / 16)        
         space = lenr / 8
-        starts = [0, startidx, startidx + space * 1, startidx + space * 2, startidx + space * 3,
-             startidx + space * 4, startidx + space * 5, startidx + space * 6, startidx + space * 7]
+        starts = int([0, startidx, startidx + space * 1, startidx + space * 2, startidx + space * 3,
+             startidx + space * 4, startidx + space * 5, startidx + space * 6, startidx + space * 7])
         if i == 7:
             start = starts[7]
             end = start + space
